Authorization

+++ backend/routes/posts.js
@@ -30,11 +30,17 @@ router.put("/:id", checkAuth, (req, res) => {
     const post = new Post({
         _id: req.body.id,
         title: req.body.title,
-        content: req.body.content
+        content: req.body.content,
+        creator: req.userData.userId
     });
-    Post.updateOne({ _id: req.params.id }, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
         .then(result => {
-            res.status(200).json({ message: "Update Successful" });
+            if (result.nModified > 0) {
+                res.status(200).json({ message: "Update Successful" });
+            }
+            else {
+                res.status(401).json({ message: "Not Authorized" });
+            }
         })
 })
 
@@ -42,7 +48,8 @@ router.post("", checkAuth, (req, res, next) => {
 
     const post = new Post({
         title: req.body.title,
-        content: req.body.content
+        content: req.body.content,
+        creator: req.userData.userId
     });
     post.save().then(createdPost => {
         res.status(201).json({
@@ -55,9 +62,13 @@ router.post("", checkAuth, (req, res, next) => {
 
 
 router.delete("/:id", checkAuth, (req, res) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Post Deleted" });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.n > 0) {
+            res.status(200).json({ message: "Post Deleted" });
+        }
+        else {
+            res.status(401).json({ message: "Not Authorized" });
+        }
     });
 });
